Create separate files for each class
