+++ src/BonusPlayer.js
@@ -1,0 +1,13 @@
+import Player from './Player.js'
+
+class BonusPlayer extends Player {
+  constructor() {
+    super(name);
+    bonusGuess() {
+      // guess 3 consonants
+      // guess 1 vowel
+    }
+  }
+}
+
+export default BonusPlayer;
