+++ src/Puzzle.js
@@ -1,0 +1,16 @@
+class Puzzle {
+  constructor() {
+    this.category = null;
+    this.answer = null;
+  }
+  randomizeCategory() {
+    // let rndCategory = random number;
+    this.category = rndCategory;
+  }
+  randomizeAnswer() {
+    // let randomAnswer = random number;
+    this.answer = randomAnswer;
+  }
+};
+
+export default Puzzle;
